Add DefaultMessagingProtocol.RelationQuery
A RelationQuery(relationName) returns the entire Relation with the
corresponding name from a Dactor if it exists. There is a specific
success message type for this case. In case of failure, e.g. no relation
of the given name existing with the Dactor, a akka.actor.Status.Failure
is returned.

This functionality can be used for tests and benchmarks and should not
necessarily be user-facing. The same is true, as we have discussed at
some point, for the DefaultMessagingProtocol.InsertIntoRelation
functionality which is why I think this is OK.